Fix carthage building, swift 2.2 support
