+++ Koloda/KolodaView/DraggableCardView/DraggableCardView.swift
@@ -1,6 +1,6 @@
 //
 //  DraggableCardView.swift
-//  TinderCardsSwift
+//  Koloda
 //
 //  Created by Eugene Andreyev on 4/23/15.
 //  Copyright (c) 2015 Yalantis. All rights reserved.
@@ -11,11 +11,11 @@ import pop
 
 protocol DraggableCardDelegate: class {
     
-    func cardDraggedWithFinishPercent(card: DraggableCardView, percent: CGFloat, direction: SwipeResultDirection)
-    func cardSwippedInDirection(card: DraggableCardView, direction: SwipeResultDirection)
-    func cardWasReset(card: DraggableCardView)
-    func cardTapped(card: DraggableCardView)
-    
+    func card(card: DraggableCardView, wasDraggedWithFinishPercent percent: CGFloat, inDirection direction: SwipeResultDirection)
+    func card(card: DraggableCardView, wasSwipedInDirection direction: SwipeResultDirection)
+    func card(cardWasReset card: DraggableCardView)
+    func card(cardWasTapped card: DraggableCardView)
+    func card(cardSwipeThresholdMargin card: DraggableCardView) -> CGFloat?
 }
 
 //Drag animation constants
@@ -39,13 +39,10 @@ public class DraggableCardView: UIView {
     
     private var panGestureRecognizer: UIPanGestureRecognizer!
     private var tapGestureRecognizer: UITapGestureRecognizer!
-    private var originalLocation: CGPoint = CGPoint(x: 0.0, y: 0.0)
     private var animationDirection: CGFloat = 1.0
     private var dragBegin = false
-    private var xDistanceFromCenter: CGFloat = 0.0
-    private var yDistanceFromCenter: CGFloat = 0.0
+    private var dragDistance = CGPointZero
     private var actionMargin: CGFloat = 0.0
-    private var firstTouch = true
     
     //MARK: Lifecycle
     init() {
@@ -65,7 +62,7 @@ public class DraggableCardView: UIView {
     
     override public var frame: CGRect {
         didSet {
-            actionMargin = frame.size.width / 2.0
+            actionMargin = delegate?.card(cardSwipeThresholdMargin: self) ?? frame.size.width / 2.0
         }
     }
     
@@ -75,15 +72,16 @@ public class DraggableCardView: UIView {
     }
     
     private func setup() {
-        panGestureRecognizer = UIPanGestureRecognizer(target: self, action: Selector("panGestureRecognized:"))
+        panGestureRecognizer = UIPanGestureRecognizer(target: self, action: #selector(DraggableCardView.panGestureRecognized(_:)))
         addGestureRecognizer(panGestureRecognizer)
-        tapGestureRecognizer = UITapGestureRecognizer(target: self, action: Selector("tapRecognized:"))
+        tapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(DraggableCardView.tapRecognized(_:)))
         addGestureRecognizer(tapGestureRecognizer)
     }
     
     //MARK: Configurations
     func configure(view: UIView, overlayView: OverlayView?) {
         self.overlayView?.removeFromSuperview()
+        self.contentView?.removeFromSuperview()
         
         if let overlay = overlayView {
             self.overlayView = overlay
@@ -95,7 +93,6 @@ public class DraggableCardView: UIView {
             self.addSubview(view)
         }
         
-        self.contentView?.removeFromSuperview()
         self.contentView = view
         configureContentView()
     }
@@ -181,47 +178,45 @@ public class DraggableCardView: UIView {
         }
     }
     
-    //MARK: GestureRecozniers
-    
+    //MARK: GestureRecognizers
     func panGestureRecognized(gestureRecognizer: UIPanGestureRecognizer) {
-        xDistanceFromCenter = gestureRecognizer.translationInView(self).x
-        yDistanceFromCenter = gestureRecognizer.translationInView(self).y
+        dragDistance = gestureRecognizer.translationInView(self)
         
         let touchLocation = gestureRecognizer.locationInView(self)
         
         switch gestureRecognizer.state {
         case .Began:
-            if firstTouch {
-                originalLocation = center
-                firstTouch = false
-            }
+            
+            let firstTouchPoint = gestureRecognizer.locationInView(self)
+            let newAnchorPoint = CGPointMake(firstTouchPoint.x / bounds.width, firstTouchPoint.y / bounds.height)
+            let oldPosition = CGPoint(x: bounds.size.width * layer.anchorPoint.x, y: bounds.size.height * layer.anchorPoint.y)
+            let newPosition = CGPoint(x: bounds.size.width * newAnchorPoint.x, y: bounds.size.height * newAnchorPoint.y)
+            layer.anchorPoint = newAnchorPoint
+            layer.position = CGPoint(x: layer.position.x - oldPosition.x + newPosition.x, y: layer.position.y - oldPosition.y + newPosition.y)
+            removeAnimations()
+            
             dragBegin = true
             
             animationDirection = touchLocation.y >= frame.size.height / 2 ? -1.0 : 1.0
             
             layer.shouldRasterize = true
             
-            pop_removeAllAnimations()
             break
         case .Changed:
-            let rotationStrength = min(xDistanceFromCenter / self.frame.size.width, rotationMax)
+            let rotationStrength = min(dragDistance.x / CGRectGetWidth(frame), rotationMax)
             let rotationAngle = animationDirection * defaultRotationAngle * rotationStrength
             let scaleStrength = 1 - ((1 - scaleMin) * fabs(rotationStrength))
             let scale = max(scaleStrength, scaleMin)
+    
+            var transform = CATransform3DIdentity
+            transform = CATransform3DScale(transform, scale, scale, 1)
+            transform = CATransform3DRotate(transform, rotationAngle, 0, 0, 1)
+            transform = CATransform3DTranslate(transform, dragDistance.x, dragDistance.y, 0)
+            layer.transform = transform
             
-            layer.rasterizationScale = scale * UIScreen.mainScreen().scale
-            
-            let transform = CGAffineTransformMakeRotation(rotationAngle)
-            let scaleTransform = CGAffineTransformScale(transform, scale, scale)
-            
-            self.transform = scaleTransform
-            center = CGPoint(x: originalLocation.x + xDistanceFromCenter, y: originalLocation.y + yDistanceFromCenter)
-            
-            updateOverlayWithFinishPercent(xDistanceFromCenter / frame.size.width)
+            updateOverlayWithFinishPercent(dragDistance.x / CGRectGetWidth(frame))
             //100% - for proportion
-            var dragDirection = SwipeResultDirection.None
-            dragDirection = xDistanceFromCenter > 0 ? .Right : .Left
-            delegate?.cardDraggedWithFinishPercent(self, percent: min(fabs(xDistanceFromCenter * 100 / frame.size.width), 100), direction: dragDirection)
+            delegate?.card(self, wasDraggedWithFinishPercent: min(fabs(dragDistance.x * 100 / CGRectGetWidth(frame)), 100), inDirection: dragDirection)
             
             break
         case .Ended:
@@ -234,10 +229,14 @@ public class DraggableCardView: UIView {
     }
     
     func tapRecognized(recogznier: UITapGestureRecognizer) {
-        delegate?.cardTapped(self)
+        delegate?.card(cardWasTapped: self)
     }
     
     //MARK: Private
+    private var dragDirection: SwipeResultDirection {
+        return dragDistance.x > 0 ? .Right : .Left
+    }
+    
     private func updateOverlayWithFinishPercent(percent: CGFloat) {
         if let overlayView = self.overlayView {
             overlayView.overlayState = percent > 0.0 ? OverlayMode.Right : OverlayMode.Left
@@ -248,138 +247,108 @@ public class DraggableCardView: UIView {
     }
     
     private func swipeMadeAction() {
-        if xDistanceFromCenter > actionMargin {
-            rightAction()
-        } else if xDistanceFromCenter < -actionMargin {
-            leftAction()
+        
+        if abs(dragDistance.x) >= actionMargin {
+            swipeAction(dragDirection)
         } else {
             resetViewPositionAndTransformations()
         }
     }
     
-    private func rightAction() {
-        let finishY = originalLocation.y + yDistanceFromCenter
-        let finishPoint = CGPoint(x: CGRectGetWidth(UIScreen.mainScreen().bounds) * 2, y: finishY)
+    private func swipeAction(direction: SwipeResultDirection) {
         
-        self.overlayView?.overlayState = OverlayMode.Right
-        self.overlayView?.alpha = 1.0
-        self.delegate?.cardSwippedInDirection(self, direction: SwipeResultDirection.Right)
-        UIView.animateWithDuration(cardSwipeActionAnimationDuration,
-            delay: 0.0,
-            options: .CurveLinear,
-            animations: {
-                self.center = finishPoint
-                
-            },
-            completion: {
-                _ in
-                
-                self.dragBegin = false
-                self.removeFromSuperview()
-        })
-    }
-    
-    private func leftAction() {
-        let finishY = originalLocation.y + yDistanceFromCenter
-        let finishPoint = CGPoint(x: -CGRectGetWidth(UIScreen.mainScreen().bounds), y: finishY)
+        let screenWidth = CGRectGetWidth(UIScreen.mainScreen().bounds)
+        let translation = screenWidth + (screenWidth / 2)
+        let directionMultiplier: CGFloat = direction == .Left ? -1 : 1
+        let finishTranslation = directionMultiplier * translation
         
-        self.overlayView?.overlayState = OverlayMode.Left
-        self.overlayView?.alpha = 1.0
-        self.delegate?.cardSwippedInDirection(self, direction: SwipeResultDirection.Left)
-        UIView.animateWithDuration(cardSwipeActionAnimationDuration,
-            delay: 0.0,
-            options: .CurveLinear,
-            animations: {
-                self.center = finishPoint
-                
-            },
-            completion: {
-                _ in
-                
-                self.dragBegin = false
-                self.removeFromSuperview()
-        })
+        overlayView?.overlayState = direction == .Left ? .Left : .Right
+        overlayView?.alpha = 1.0
+        delegate?.card(self, wasSwipedInDirection: direction)
+        let translationAnimation = POPBasicAnimation(propertyNamed: kPOPLayerTranslationX)
+        translationAnimation.duration = cardSwipeActionAnimationDuration
+        translationAnimation.fromValue = POPLayerGetTranslationX(layer)
+        translationAnimation.toValue = finishTranslation
+        translationAnimation.completionBlock = { _, _ in
+            self.removeFromSuperview()
+        }
+        layer.pop_addAnimation(translationAnimation, forKey: "swipeTranslationAnimation")
     }
     
     private func resetViewPositionAndTransformations() {
-        self.delegate?.cardWasReset(self)
+        delegate?.card(cardWasReset: self)
         
-        let resetPositionAnimation = POPSpringAnimation(propertyNamed: kPOPLayerPosition)
+        removeAnimations()
         
-        resetPositionAnimation.toValue = NSValue(CGPoint: originalLocation)
+        let resetPositionAnimation = POPSpringAnimation(propertyNamed: kPOPLayerTranslationXY)
+        resetPositionAnimation.fromValue = NSValue(CGPoint: CGPoint(x: dragDistance.x, y: dragDistance.y))
+        resetPositionAnimation.toValue = NSValue(CGPoint: CGPointZero)
         resetPositionAnimation.springBounciness = cardResetAnimationSpringBounciness
         resetPositionAnimation.springSpeed = cardResetAnimationSpringSpeed
         resetPositionAnimation.completionBlock = {
             (_, _) in
-            
+            self.layer.transform = CATransform3DIdentity
             self.dragBegin = false
         }
         
-        pop_addAnimation(resetPositionAnimation, forKey: cardResetAnimationKey)
+        layer.pop_addAnimation(resetPositionAnimation, forKey: "resetPositionAnimation")
         
-        UIView.animateWithDuration(cardResetAnimationDuration,
-            delay: 0.0,
-            options: [.CurveLinear, .AllowUserInteraction],
-            animations: {
-                self.transform = CGAffineTransformMakeRotation(0)
-                self.overlayView?.alpha = 0
-                self.layoutIfNeeded()
-                
-                return
-            },
-            completion: {
-                _ in
-                
-                self.transform = CGAffineTransformIdentity
-                
-                return
-        })
+        let resetRotationAnimation = POPBasicAnimation(propertyNamed: kPOPLayerRotation)
+        resetRotationAnimation.fromValue = POPLayerGetRotationZ(layer)
+        resetRotationAnimation.toValue = CGFloat(0.0)
+        resetRotationAnimation.duration = cardResetAnimationDuration
+        
+        layer.pop_addAnimation(resetRotationAnimation, forKey: "resetRotationAnimation")
+        
+        let overlayAlphaAnimation = POPBasicAnimation(propertyNamed: kPOPViewAlpha)
+        overlayAlphaAnimation.toValue = 0.0
+        overlayAlphaAnimation.duration = cardResetAnimationDuration
+        overlayAlphaAnimation.completionBlock = { _, _ in
+            self.overlayView?.alpha = 0
+        }
+        overlayView?.pop_addAnimation(overlayAlphaAnimation, forKey: "resetOverlayAnimation")
+        
+        let resetScaleAnimation = POPBasicAnimation(propertyNamed: kPOPLayerScaleXY)
+        resetScaleAnimation.toValue = NSValue(CGPoint: CGPoint(x: 1.0, y: 1.0))
+        resetScaleAnimation.duration = cardResetAnimationDuration
+        layer.pop_addAnimation(resetScaleAnimation, forKey: "resetScaleAnimation")
     }
     
     //MARK: Public
-    
-    func swipeLeft () {
-        if !dragBegin {
-            
-            let finishPoint = CGPoint(x: -CGRectGetWidth(UIScreen.mainScreen().bounds), y: center.y)
-            self.delegate?.cardSwippedInDirection(self, direction: SwipeResultDirection.Left)
-            UIView.animateWithDuration(cardSwipeActionAnimationDuration,
-                delay: 0.0,
-                options: .CurveLinear,
-                animations: {
-                    self.center = finishPoint
-                    self.transform = CGAffineTransformMakeRotation(CGFloat(-M_PI_4))
-                    
-                    return
-                },
-                completion: {
-                    _ in
-                    
-                    self.removeFromSuperview()
-                    
-                    return
-            })
-        }
+    func removeAnimations() {
+        pop_removeAllAnimations()
+        layer.pop_removeAllAnimations()
     }
     
-    func swipeRight () {
+    func swipe(direction: SwipeResultDirection) {
         if !dragBegin {
+            delegate?.card(self, wasSwipedInDirection: direction)
             
-            let finishPoint = CGPoint(x: CGRectGetWidth(UIScreen.mainScreen().bounds) * 2, y: center.y)
-            self.delegate?.cardSwippedInDirection(self, direction: SwipeResultDirection.Right)
-            UIView.animateWithDuration(cardSwipeActionAnimationDuration, delay: 0.0, options: .CurveLinear, animations: {
-                    self.center = finishPoint
-                    self.transform = CGAffineTransformMakeRotation(CGFloat(M_PI_4))
-                    
-                    return
-                },
-                completion: {
-                    _ in
-                    
-                    self.removeFromSuperview()
-                    
-                    return
-            })
+            let screenWidth = CGRectGetWidth(UIScreen.mainScreen().bounds)
+            let finalPosition = direction == .Left ? -screenWidth : 2 * screenWidth
+            
+            let swipePositionAnimation = POPBasicAnimation(propertyNamed: kPOPLayerPositionX)
+            swipePositionAnimation.toValue = finalPosition
+            swipePositionAnimation.duration = cardSwipeActionAnimationDuration
+            swipePositionAnimation.completionBlock = {
+                (_, _) in
+                self.removeFromSuperview()
+            }
+            
+            layer.pop_addAnimation(swipePositionAnimation, forKey: "swipePositionAnimation")
+            
+            let swipeRotationAnimation = POPBasicAnimation(propertyNamed: kPOPLayerRotation)
+            swipeRotationAnimation.fromValue = POPLayerGetRotationZ(layer)
+            swipeRotationAnimation.toValue = CGFloat(direction == .Left ? -M_PI_4 : M_PI_4)
+            swipeRotationAnimation.duration = cardSwipeActionAnimationDuration
+            
+            layer.pop_addAnimation(swipeRotationAnimation, forKey: "swipeRotationAnimation")
+            
+            overlayView?.overlayState = direction == .Left ? .Left : .Right
+            let overlayAlphaAnimation = POPBasicAnimation(propertyNamed: kPOPViewAlpha)
+            overlayAlphaAnimation.toValue = 1.0
+            overlayAlphaAnimation.duration = cardSwipeActionAnimationDuration
+            overlayView?.pop_addAnimation(overlayAlphaAnimation, forKey: "swipeOverlayAnimation")
         }
     }
 }
