+++ Koloda/KolodaView/OverlayView/OverlayView.swift
@@ -1,13 +1,12 @@
 //
 //  OverlayView.swift
-//  TinderCardsSwift
+//  Koloda
 //
 //  Created by Eugene Andreyev on 4/24/15.
 //  Copyright (c) 2015 Eugene Andreyev. All rights reserved.
 //
 
 import UIKit
-import pop
 
 public enum OverlayMode{
     case None
